requirements.txt
PyPDF2
nltk
scikit-learn
pandas
numpy
python-dotenv
pymongo